+++ R/storage_utils.R
@@ -1,38 +1,63 @@
-do_container_op <- function(container, path="", options=list(), headers=list(), http_verb="GET", ...)
+#' Carry out operations on a storage account container or endpoint
+#'
+#' @param container,endpoint For `do_container_op`, a storage container object (inheriting from `storage_container`). For `call_storage_endpoint`, a  storage endpoint object (inheriting from `storage_endpoint`).
+#' @param operation The container operation to perform, which will form part of the URL path.
+#' @param path The path component of the endpoint call.
+#' @param options A named list giving the query parameters for the operation.
+#' @param headers A named list giving any additional HTTP headers to send to the host. Note that AzureStor will handle authentication details, so you don't have to specify these here.
+#' @param body The request body for a `PUT/POST/PATCH` call.
+#' @param ... Any additional arguments to pass to `httr::VERB`.
+#' @param http_verb The HTTP verb as a string, one of `GET`, `DELETE`, `PUT`, `POST`, `HEAD` or `PATCH`.
+#' @param http_status_handler The R handler for the HTTP status code of the response. `"stop"`, `"warn"` or `"message"` will call the corresponding handlers in httr, while `"pass"` ignores the status code. The latter is primarily useful for debugging purposes.
+#' @param progress Used by the file transfer functions, to display a progress bar.
+#' @details
+#' These functions form the low-level interface between R and the storage API. `do_container_op` constructs a path from the operation and the container name, and passes it and the other arguments to `call_storage_endpoint`.
+#' @return
+#' If `http_status_handler` is one of `"stop"`, `"warn"` or `"message"`, the status code of the response is checked, and if an error is not thrown, the parsed content of the response is returned. An exception is if the response was written to disk, as part of a file download; in this case, the return value is NULL.
+#'
+#' If `http_status_handler` is `"pass"`, the entire response is returned without modification.
+#' @seealso
+#' [blob_endpoint], [file_endpoint], [adls_endpoint]
+#'
+#' [blob_container], [file_share], [adls_filesystem]
+#'
+#' [httr::GET], [httr::PUT], [httr::POST], [httr::PATCH], [httr::HEAD], [httr::DELETE]
+#' @rdname storage_call
+#' @export
+do_container_op <- function(container, operation="", options=list(), headers=list(), http_verb="GET", ...)
 {
     endp <- container$endpoint
 
     # don't add trailing / if no within-container path supplied: ADLS will complain
-    path <- if(nchar(path) > 0)
-        sub("//", "/", paste0(container$name, "/", path))
+    operation <- if(nchar(operation) > 0)
+        sub("//", "/", paste0(container$name, "/", operation))
     else container$name
 
-    invisible(do_storage_call(endp$url, path, options=options, headers=headers,
-                              key=endp$key, token=endp$token, sas=endp$sas, api_version=endp$api_version,
+    invisible(call_storage_endpoint(endp, operation, options=options, headers=headers,
                               http_verb=http_verb, ...))
 }
 
 
-do_storage_call <- function(endpoint_url, path, options=list(), headers=list(), body=NULL, ...,
-                            key=NULL, token=NULL, sas=NULL,
-                            api_version=getOption("azure_storage_api_version"),
-                            http_verb=c("GET", "DELETE", "PUT", "POST", "HEAD", "PATCH"),
-                            http_status_handler=c("stop", "warn", "message", "pass"),
-                            progress=NULL)
+#' @rdname storage_call
+#' @export
+call_storage_endpoint <- function(endpoint, path, options=list(), headers=list(), body=NULL, ...,
+                                  http_verb=c("GET", "DELETE", "PUT", "POST", "HEAD", "PATCH"),
+                                  http_status_handler=c("stop", "warn", "message", "pass"),
+                                  progress=NULL)
 {
     verb <- match.arg(http_verb)
-    url <- httr::parse_url(endpoint_url)
+    url <- httr::parse_url(endpoint$url)
     url$path <- URLencode(path)
     if(!is_empty(options))
         url$query <- options[order(names(options))] # must be sorted for access key signing
 
     # use key if provided, otherwise AAD token if provided, otherwise sas if provided, otherwise anonymous access
-    if(!is.null(key))
-        headers <- sign_request(key, verb, url, headers, api_version)
-    else if(!is.null(token))
-        headers <- add_token(token, headers, api_version)
-    else if(!is.null(sas))
-        url <- add_sas(sas, url)
+    if(!is.null(endpoint$key))
+        headers <- sign_request(endpoint$key, verb, url, headers, endpoint$api_version)
+    else if(!is.null(endpoint$token))
+        headers <- add_token(endpoint$token, headers, endpoint$api_version)
+    else if(!is.null(endpoint$sas))
+        url <- add_sas(endpoint$sas, url)
 
     headers <- do.call(httr::add_headers, headers)
     retries <- as.numeric(getOption("azure_storage_retries"))