+++ tests/testthat/test04_adls.R
@@ -127,27 +127,6 @@ test_that("ADLSgen2 client interface works",
     iris3 <- as.data.frame(jsonlite::fromJSON(con))
     expect_identical(iris, iris3)
 
-    # multiple file transfers
-    files <- lapply(1:10, function(f) paste0(sample(letters, 1000, replace=TRUE), collapse=""))
-    filenames <- sapply(1:10, function(n) file.path(tempdir(), sprintf("multitransfer_%d", n)))
-    suppressWarnings(file.remove(filenames))
-    mapply(writeLines, files, filenames)
-
-    create_adls_dir(fs, "multi")
-    multiupload_adls_file(fs, file.path(tempdir(), "multitransfer_*"), "multi")
-
-    dest_dir <- file.path(tempdir(), "adls_multitransfer")
-    suppressWarnings(unlink(dest_dir, recursive=TRUE))
-    dir.create(dest_dir)
-    multidownload_adls_file(fs, "multi/multitransfer_*", dest_dir, overwrite=TRUE)
-
-    expect_true(all(sapply(filenames, function(f)
-    {
-        src <- readBin(f, "raw", n=1e5)
-        dest <- readBin(file.path(dest_dir, basename(f)), "raw", n=1e5)
-        identical(src, dest)
-    })))
-
     # ways of deleting a filesystem
     delete_adls_filesystem(fs, confirm=FALSE)
     delete_adls_filesystem(ad, "newfs2", confirm=FALSE)
@@ -173,26 +152,6 @@ test_that("AAD authentication works",
     suppressWarnings(file.remove(tok_dl))
     download_adls_file(fs, "iris.csv", tok_dl)
     expect_identical(readBin(orig_file, "raw", n=1e5), readBin(tok_dl, "raw", n=1e5))
-
-    # multiple upload and download
-    files <- lapply(1:10, function(f) paste0(sample(letters, 1000, replace=TRUE), collapse=""))
-    filenames <- sapply(1:10, function(n) file.path(tempdir(), sprintf("multitransfer_%d", n)))
-    suppressWarnings(file.remove(filenames))
-    mapply(writeLines, files, filenames)
-
-    multiupload_adls_file(fs, file.path(tempdir(), "multitransfer_*"), "/")
-
-    dest_dir <- file.path(tempdir(), "adls_multitransfer")
-    suppressWarnings(unlink(dest_dir, recursive=TRUE))
-    dir.create(dest_dir)
-    multidownload_adls_file(fs, "multitransfer_*", dest_dir, overwrite=TRUE)
-
-    expect_true(all(sapply(filenames, function(f)
-    {
-        src <- readBin(f, "raw", n=1e5)
-        dest <- readBin(file.path(dest_dir, basename(f)), "raw", n=1e5)
-        identical(src, dest)
-    })))
 
     delete_adls_filesystem(fs, confirm=FALSE)
     expect_true(is_empty(list_adls_filesystems(ad)))
@@ -293,59 +252,6 @@ test_that("chunked downloading works",
 
     con <- download_adls_file(fs, "iris.csv", NULL, blocksize=150)
     expect_identical(readBin(orig_file, "raw", n=1e5), readBin(con, "raw", n=1e5))
-})
-
-
-test_that("vector source for upload/download works",
-{
-    write_file <- function(fname)
-    {
-        bytes <- openssl::rand_bytes(1000)
-        writeBin(bytes, file.path(srcdir, fname))
-        invisible(fname)
-    }
-
-    ad <- stor$get_adls_endpoint()
-    fs <- create_adls_filesystem(ad, "vectransfer")
-
-    srcdir <- tempfile()
-    destdir <- tempfile()
-    destdir2 <- tempfile()
-    destdir3 <- tempfile()
-    dir.create(srcdir)
-    dir.create(destdir)
-    dir.create(destdir2)
-    dir.create(destdir3)
-
-    srcs <- unlist(lapply(letters[1:3], function(letter)
-    {
-        for(i in 1:4)
-            write_file(paste0(letter, i, collapse=""))
-        paste0(letter, "*")
-    }))
-    srcs <- c(srcs, write_file("d1"), write_file("d2"))
-
-    multiupload_adls_file(fs, file.path(srcdir, srcs), "/")
-    multidownload_adls_file(fs, srcs, destdir)
-
-    expect_identical(pool_size(), 10L)
-    expect_identical(dir(srcdir), dir(destdir))
-
-    create_adls_dir(fs, "/newdir")
-    multiupload_adls_file(fs, file.path(srcdir, srcs), "/newdir")
-    multidownload_adls_file(fs, file.path("/newdir", srcs), destdir2)
-
-    expect_identical(pool_size(), 10L)
-    expect_identical(dir(srcdir), dir(destdir2))
-
-    srcs <- dir(srcdir)
-    dests <- paste0("destname_", srcs)
-    expect_identical(length(srcs), length(dests))
-    multiupload_adls_file(fs, file.path(srcdir, srcs), dests)
-    multidownload_adls_file(fs, dests, destdir3)
-
-    expect_identical(pool_size(), 10L)
-    expect_identical(dests, dir(destdir3))
 })